requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==2.5.1
python-dotenv==1.0.0
